\name{extractDataFromIgraph}
\alias{extractDataFromIgraph}
\title{extractDataFromIgraph}
\usage{
extractDataFromIgraph(f, graph)
}
\arguments{
  \item{f}{formula}

  \item{graph}{An igraph object}
}
\value{
A list of data
}
\description{
extract data from igraph object using formula
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\keyword{internal}

