\name{genData}
\alias{genData}
\title{genData}
\usage{
genData(network_size = 50, beta = 1:3, theta = beta, lambda = 0.3,
  degree_dist = 1:5, degree_prob = rep(1/length(degree_dist),
  length(degree_dist)))
}
\arguments{
  \item{network_size}{A vector of the size of networks}

  \item{beta}{Own effect}

  \item{theta}{Contextual Effect}

  \item{lambda}{Spillover Effect}

  \item{degree_dist}{value of possible degree}

  \item{degree_prob}{probability distribution of the
  degree}
}
\value{
value A list of igraph object
}
\description{
genData create a list of igraph objects. The vertices of
each graph containts properties of variable y and variables
x which are generated by social interaction model.
Mathematically, Y = X \\beta + WX \\theta + WY \\lambda + e
}
\examples{
\dontrun{
  graph = genData(network_size = c(50,50),
                  beta = (1:3)*10,
                  theta=(1:3)*10,
                  lambda=0.3,
                  degree_dist=2:3,
                  degree_prob=c(0.5,0.5) )
  socInt(f=y ~ x1 + x2 + x3 | x1 + x2 + x3, graph)
}
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

