\name{generateWeighting}
\alias{generateWeighting}
\title{generateWeighting}
\usage{
generateWeighting(D)
}
\arguments{
  \item{D}{a network matrix}
}
\value{
weighting matrix
}
\description{
Generate weighting matrix from adjacency matrix.
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

