\name{socInt.maxLik}
\alias{socInt.maxLik}
\title{socInt.maxLik}
\usage{
socInt.maxLik(y, x, wy, n, W)
}
\arguments{
  \item{y}{vector of outcome variables}

  \item{x}{vector of rhs variables}

  \item{wy}{vector of weighted peers outcome}

  \item{n}{vector of network sizes}

  \item{W}{A list of network weighting matrix}
}
\value{
value maxLik Object
}
\description{
Estimate social interaction model with maximum likelihood
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

